+++ src/InvoiceTypeCode.php
@@ -9,7 +9,9 @@ namespace NumNum\UBL;
 class InvoiceTypeCode
 {
     const INVOICE = 380;
-    const CORRECTED_INVOICE = 384;
-    const SELF_BILLING_INVOICE = 389;
     const CREDIT_NOTE = 381;
+    const DEBIT_NOTE = 383;
+    const CORRECTED_INVOICE = 384;
+    const ADVANCE_INVOICE = 386;
+    const SELF_BILLING_INVOICE = 389;
 }