+++ src/LegalMonetaryTotal.php
@@ -11,6 +11,7 @@ class LegalMonetaryTotal implements XmlSerializable
     private $taxExclusiveAmount = 0;
     private $taxInclusiveAmount = 0;
     private $allowanceTotalAmount = 0;
+    private $prepaidAmount;
     private $payableAmount = 0;
 
     /**
@@ -86,6 +87,24 @@ class LegalMonetaryTotal implements XmlSerializable
     }
 
     /**
+     * @return ?float
+     */
+    public function getPrepaidAmount(): ?float
+    {
+        return $this->prepaidAmount;
+    }
+
+    /**
+     * @param ?float $prepaidAmount
+     * @return LegalMonetaryTotal
+     */
+    public function setPrepaidAmount(?float $prepaidAmount): LegalMonetaryTotal
+    {
+        $this->prepaidAmount = $prepaidAmount;
+        return $this;
+    }
+
+    /**
      * @return float
      */
     public function getPayableAmount(): ?float
@@ -143,7 +162,22 @@ class LegalMonetaryTotal implements XmlSerializable
                     'currencyID' => Generator::$currencyID
                 ]
 
-            ],
+            ]
+        ]);
+
+        if ($this->prepaidAmount !== null) {
+            $writer->write([
+                [
+                    'name' => Schema::CBC . 'PrepaidAmount',
+                    'value' => number_format($this->prepaidAmount, 2, '.', ''),
+                    'attributes' => [
+                        'currencyID' => Generator::$currencyID
+                    ]
+                ]
+            ])
+        }
+
+        $writer->write([
             [
                 'name' => Schema::CBC . 'PayableAmount',
                 'value' => number_format($this->payableAmount, 2, '.', ''),